Resolve clang-tidy warnings=error; Notable std::vector<>: size() -> size_t (unsigned) but iterator and difference_type are signed and losing half of range
Because std::vector<>::begin() iterator performs arithmetic
using a signed difference_type, we need to use such a signed type
here to avoid `bugprone-narrowing-conversions` (LINT)

Now, isn't this odd as std::vector<>::size() uses unsigned size_type,
aka size_t and mentioned iterator hence lose half the value range possible?

+++ src/lesson02b_arithmetic.cpp
@@ -47,8 +47,8 @@ long summe_loop( long a, long b){
  * @return
  */
 long summe_gaus( long a, long b ){
-   double n = b - a + 1;
-   double r = ( a + b ) * ( n / 2.0 );
+   double n = static_cast<double>(b - a + 1);
+   double r = static_cast<double>( a + b ) * ( n / 2.0 );
    return (long)r;
 }
